functions/upload_media/requirements.txt
appwrite==9.0.3
asgiref==3.8.1
astroid==3.3.11
black==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.2.1
dill==0.4.0
idna==3.10
isort==6.0.1
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pylint==3.3.7
requests==2.32.3
sqlparse==0.5.3
tomlkit==0.13.3
urllib3==2.4.0
